+++ UI/Controllers/TelegramBotController.cs
@@ -11,13 +11,16 @@ public class TelegramBotController : BaseController<TelegramBotController>
 {
     private readonly ITelegramBotService _telegramBotService;
     private readonly ITelegramService _telegramService;
+    private readonly IVerifyService _verifyService;
 
     public TelegramBotController(ITelegramBotService telegramBotService, ITelegramService telegramService,
+        IVerifyService verifyService,
         ILogger<TelegramBotController> logger) :
         base(logger)
     {
         _telegramBotService = telegramBotService;
         _telegramService = telegramService;
+        _verifyService = verifyService;
     }
 
     [HttpGet]
@@ -33,13 +36,18 @@ public class TelegramBotController : BaseController<TelegramBotController>
     {
         try
         {
+            var available = await _telegramService.IsBotAvailableAsync(telegramBot.Token).ConfigureAwait(false);
+
+            if (!available) return BadRequest(JsonConvert.SerializeObject("Bot is not available"));
+
+            if (telegramBot.IsActive) await _telegramService.SetWebHookAsync(telegramBot.Token).ConfigureAwait(false);
 
             await _telegramBotService.CreateTelegramBotAsync(User.Claims, telegramBot).ConfigureAwait(false);
             return Ok();
         }
         catch (Exception ex)
         {
-            Logger.LogError(ex, "Error while updating telegram bot");
+            Logger.LogError(ex, "Error while creating telegram bot");
             return BadRequest(ex.GetErrorMessageJson());
         }
     }
@@ -49,6 +57,17 @@ public class TelegramBotController : BaseController<TelegramBotController>
     {
         try
         {
+            if (!await _verifyService.VerifyTelegramBotAsync(User.Claims, id).ConfigureAwait(false)) return NotFound();
+
+            var available = await _telegramService.IsBotAvailableAsync(telegramBot.Token).ConfigureAwait(false);
+
+            if (!available) return BadRequest(JsonConvert.SerializeObject("Bot is not available"));
+
+            if (telegramBot.IsActive)
+                await _telegramService.SetWebHookAsync(telegramBot.Token).ConfigureAwait(false);
+            else
+                await _telegramService.DeleteWebHookAsync(telegramBot.Token).ConfigureAwait(false);
+
             await _telegramBotService.UpdateTelegramBotAsync(id, telegramBot).ConfigureAwait(false);
             return Ok();
         }
@@ -64,10 +83,14 @@ public class TelegramBotController : BaseController<TelegramBotController>
     {
         try
         {
+            if (!await _verifyService.VerifyTelegramBotAsync(User.Claims, id).ConfigureAwait(false)) return NotFound();
+
+            var bot = await _telegramBotService.GetTelegramBotAsync(id).ConfigureAwait(false);
+            await _telegramService.DeleteWebHookAsync(bot.Token).ConfigureAwait(false);
             await _telegramBotService.DeleteTelegramBotAsync(id).ConfigureAwait(false);
             return Ok();
         }
-        catch(Exception ex)
+        catch (Exception ex)
         {
             Logger.LogError(ex, "Error while deleting telegram bot");
             return BadRequest(ex.GetErrorMessageJson());
